refactor: 코드 디렉토리 변경
