+++ src/main/java/domain/game/PlayerGameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public record PlayerGameResult(String name, int profit) {
 }
