+++ src/main/java/domain/participant/Hand.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import domain.card.Card;
 import java.util.ArrayList;
