+++ src/main/java/domain/participant/Name.java
@@ -1,14 +1,12 @@
-package domain;
+package domain.participant;
 
-public class Name {
+public record Name(String name) {
     public static final int NAME_LENGTH_THRESHOLD = 5;
     public static final String ERROR_NAME_EMPTY = "[ERROR] 이름은 비어있을 수 없습니다.";
     public static final String ERROR_NAME_LENGTH = "[ERROR] 이름은 5자를 초과할 수 없습니다.";
-    private final String name;
 
-    public Name(String name) {
+    public Name {
         validateNameLength(name);
-        this.name = name;
     }
 
     private void validateNameLength(String name) {
@@ -19,9 +17,5 @@ public class Name {
         if (name.length() > NAME_LENGTH_THRESHOLD) {
             throw new IllegalArgumentException(ERROR_NAME_LENGTH);
         }
-    }
-
-    public String getName() {
-        return name;
     }
 }
