+++ src/test/java/domain/BettingTest.java
@@ -11,6 +11,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import domain.card.Card;
 import domain.card.Rank;
+import domain.game.Betting;
+import domain.game.WinningStatus;
 import domain.participant.Dealer;
 import domain.participant.Player;
 import org.junit.jupiter.api.Test;
