+++ src/test/java/domain/strategy/ShuffleStrategyTest.java
@@ -2,7 +2,7 @@ package domain.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.Deck;
+import domain.card.Deck;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Suit;
